simplify refactor + remove tests
