+++ pkg/process/checks/process_nix_test.go
@@ -9,27 +9,14 @@ package checks
 
 import (
 	"fmt"
-	"math/rand/v2"
 	"regexp"
 	"sort"
-	"strings"
 	"testing"
 	"time"
 
-	wmdef "github.com/DataDog/datadog-agent/comp/core/workloadmeta/def"
-	wmimpl "github.com/DataDog/datadog-agent/comp/core/workloadmeta/impl"
-	gpusubscriberfxmock "github.com/DataDog/datadog-agent/comp/process/gpusubscriber/fx-mock"
-	configmock "github.com/DataDog/datadog-agent/pkg/config/mock"
-	"github.com/DataDog/datadog-agent/pkg/process/metadata"
-	probeMocks "github.com/DataDog/datadog-agent/pkg/process/procutil/mocks"
-	"github.com/DataDog/datadog-go/v5/statsd"
-	"github.com/benbjohnson/clock"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
-	"github.com/stretchr/testify/require"
-
 	model "github.com/DataDog/agent-payload/v5/process"
 	"github.com/shirou/gopsutil/v4/cpu"
+	"github.com/stretchr/testify/assert"
 
 	"github.com/DataDog/datadog-agent/pkg/process/metadata/parser"
 	"github.com/DataDog/datadog-agent/pkg/process/procutil"
@@ -440,695 +427,5 @@ func TestProcessGPUTagging(t *testing.T) {
 				assert.Equal(t, proc.Tags, tc.pidToGPUTags[proc.Pid])
 			}
 		})
-	}
-}
-
-func processCheckWithMockProbeWLM(t *testing.T, elevatedSystemProbePermissions bool) (*ProcessCheck, *probeMocks.Probe, *wmimpl.MockWLM) {
-	t.Helper()
-	probe := probeMocks.NewProbe(t)
-	sysInfo := &model.SystemInfo{
-		Cpus: []*model.CPUInfo{
-			{CoreId: "1"},
-			{CoreId: "2"},
-			{CoreId: "3"},
-			{CoreId: "4"},
-		},
-	}
-	hostInfo := &HostInfo{
-		SystemInfo: sysInfo,
-	}
-	serviceExtractorEnabled := true
-	useWindowsServiceName := true
-	useImprovedAlgorithm := false
-	serviceExtractor := parser.NewServiceExtractor(serviceExtractorEnabled, useWindowsServiceName, useImprovedAlgorithm)
-
-	mockGpuSubscriber := gpusubscriberfxmock.SetupMockGpuSubscriber(t)
-
-	// Workload meta + tagger
-	mockWLM := wmimpl.NewMockWLM(t)
-
-	return &ProcessCheck{
-		wmeta:                   mockWLM,
-		useWLMProcessCollection: true,
-		probe:                   probe,
-		scrubber:                procutil.NewDefaultDataScrubber(),
-		hostInfo:                hostInfo,
-		containerProvider:       mockContainerProvider(t),
-		sysProbeConfig: &SysProbeConfig{
-			ProcessModuleEnabled: elevatedSystemProbePermissions,
-		},
-		checkCount:       0,
-		skipAmount:       2,
-		serviceExtractor: serviceExtractor,
-		extractors:       []metadata.Extractor{serviceExtractor},
-		gpuSubscriber:    mockGpuSubscriber,
-		statsd:           &statsd.NoOpClient{},
-		timer:            clock.New(),
-	}, probe, mockWLM
-}
-
-func wlmProcessWithCreateTime(pid int32, spaceSeparatedCmdline string, creationTime int64) *wmdef.Process {
-	return &wmdef.Process{
-		EntityID: wmdef.EntityID{
-			ID:   fmt.Sprintf("%d", pid),
-			Kind: wmdef.KindProcess,
-		},
-		Pid:          pid,
-		Cmdline:      strings.Split(spaceSeparatedCmdline, " "),
-		CreationTime: time.Unix(creationTime, 0),
-	}
-}
-
-func procutilCPUTimeToCPUTimeStat(t *procutil.CPUTimesStat) cpu.TimesStat {
-	return cpu.TimesStat{
-		CPU:       "cpu-total", // hardcoded to cpu-total when percpu is False - cpuTimer.Times(false)
-		User:      t.User,
-		System:    t.System,
-		Idle:      t.Idle,
-		Nice:      t.Nice,
-		Iowait:    t.Iowait,
-		Irq:       t.Irq,
-		Softirq:   t.Softirq,
-		Steal:     t.Steal,
-		Guest:     t.Guest,
-		GuestNice: t.GuestNice,
-		// Other fields (like Stolen and Timestamp) are not in TimesStat
-	}
-}
-
-func modelProcFromWLMProc(wlmProc *wmdef.Process, stats *procutil.Stats, lastStats *procutil.Stats, processContext []string, now time.Time, lastRunTime time.Time) *model.Process {
-
-	return &model.Process{
-		Pid:   wlmProc.Pid,
-		NsPid: wlmProc.NsPid,
-		//Host:                   nil, // not set
-		Command: &model.Command{
-			Args: wlmProc.Cmdline,
-			Cwd:  wlmProc.Cwd,
-			//Root:   "", // not used
-			//OnDisk: false, // not used
-			Ppid: wlmProc.Ppid,
-			//Pgroup: 0, // not used
-			Exe:  wlmProc.Exe,
-			Comm: wlmProc.Comm,
-		},
-		User:   &model.ProcessUser{},
-		Memory: formatMemory(stats),
-		// TODO: test when cpu time is nil
-		Cpu:        formatCPU(stats, lastStats, procutilCPUTimeToCPUTimeStat(stats.CPUTime), procutilCPUTimeToCPUTimeStat(lastStats.CPUTime)),
-		CreateTime: wlmProc.CreationTime.Unix(), // process check uses stats.create_time but this should be equal to wlmProc.creation_time
-		//Container:              nil, // TODO: container data not yet queried from wlm
-		OpenFdCount: stats.OpenFdCount,
-		State:       model.ProcessState(model.ProcessState_value[stats.Status]),
-		IoStat:      formatIO(stats, lastStats.IOStat, now, lastRunTime),
-		//ContainerId:            "", // TODO: container data not yet queried from wlm
-		//ContainerKey:           0, // TODO: container data not yet queried from wlm
-		VoluntaryCtxSwitches:   uint64(stats.CtxSwitches.Voluntary),
-		InvoluntaryCtxSwitches: uint64(stats.CtxSwitches.Involuntary),
-		ByteKey:                nil, // not used
-		ContainerByteKey:       nil, // not used
-		Networks:               nil, // not used
-		ProcessContext:         processContext,
-		Tags:                   nil,
-	}
-}
-
-// randRange returns a random number between min and max inclusive [min, max]
-func randRange(min, max int) int {
-	return rand.IntN(max+1-min) + min
-}
-
-// randomUnprivilegedProcessStats returns process stats with reasonable randomized data
-func randomProcessStats(createTime int64, withPriviledgedData bool) *procutil.Stats {
-	proc := &procutil.Stats{CreateTime: createTime, // 1 second to 1 day ago
-		Status:     []string{"U", "D", "R", "S", "T", "W", "Z"}[rand.IntN(7)], // Valid process statuses
-		Nice:       int32(randRange(-20, 19)),                                 // -20 (highest priority) to 19 (lowest)
-		NumThreads: int32(randRange(1, 500)),                                  // Most processes use <100 threads, upper bound for heavy apps
-
-		CPUPercent: &procutil.CPUPercentStat{
-			UserPct:   rand.Float64() * 100, // Simulate 0-100% user CPU usage
-			SystemPct: rand.Float64() * 100, // Simulate 0-100% system CPU usage
-		},
-
-		CPUTime: &procutil.CPUTimesStat{
-			User:      rand.Float64() * 10000, // Seconds spent in user mode
-			System:    rand.Float64() * 5000,  // Seconds spent in kernel mode
-			Idle:      rand.Float64() * 20000, // Idle time on thread pools or waiting
-			Nice:      rand.Float64() * 1000,  // Niceness-adjusted CPU time
-			Iowait:    rand.Float64() * 1000,  // Waiting on IO
-			Irq:       rand.Float64() * 500,   // Time servicing IRQs
-			Softirq:   rand.Float64() * 500,   // Time servicing soft IRQs
-			Steal:     rand.Float64() * 100,   // Time stolen by hypervisor
-			Guest:     rand.Float64() * 50,    // Guest VM time (if applicable)
-			GuestNice: rand.Float64() * 50,    // Guest time with nice value
-			Stolen:    rand.Float64() * 10,    // Time stolen from a virtual CPU
-			Timestamp: time.Now().Unix(),      // Capture time
-		},
-
-		MemInfo: &procutil.MemoryInfoStat{
-			RSS:  uint64(randRange(1<<20, 500<<20)), // 1MB–500MB resident memory
-			VMS:  uint64(randRange(10<<20, 5<<30)),  // 10MB–5GB virtual memory
-			Swap: uint64(randRange(0, 1<<30)),       // 0–1GB swap
-		},
-
-		MemInfoEx: &procutil.MemoryInfoExStat{
-			RSS:    uint64(randRange(1<<20, 500<<20)), // should be the same as MemInfo.RSS
-			VMS:    uint64(randRange(10<<20, 5<<30)),  // should be the same as MemInfo.VMS
-			Shared: uint64(randRange(0, 100<<20)),     // Shared memory (e.g. mmap)
-			Text:   uint64(randRange(0, 50<<20)),      // Executable code
-			Lib:    uint64(randRange(0, 50<<20)),      // Shared libraries
-			Data:   uint64(randRange(1<<20, 2<<30)),   // Heap/stack/data
-			Dirty:  uint64(randRange(0, 10<<20)),      // Pages waiting to be written
-		},
-
-		CtxSwitches: &procutil.NumCtxSwitchesStat{
-			Voluntary:   int64(randRange(0, 1_000_000)), // Caused by blocking or waiting
-			Involuntary: int64(randRange(0, 500_000)),   // Caused by CPU scheduler preemption
-		},
-	}
-	if withPriviledgedData {
-		proc.IOStat = &procutil.IOCountersStat{
-			ReadCount:  int64(randRange(0, 1_000_000)), // Number of read syscalls
-			WriteCount: int64(randRange(0, 1_000_000)), // Number of write syscalls
-			ReadBytes:  int64(randRange(0, 10<<30)),    // Up to 10GB read
-			WriteBytes: int64(randRange(0, 10<<30)),    // Up to 10GB written
-		}
-		// IORateStat is never populated. TODO: we should probably remove it
-		//proc.IORateStat = &procutil.IOCountersRateStat{}
-		proc.OpenFdCount = int32(randRange(0, 5000)) // 3 minimum (stdin/out/err) to thousands for busy daemons
-	}
-	return proc
-}
-
-// createTestProcessData creates 5 test processes with associated random stats
-func createTestWLMProcessData(createTime time.Time, elevatedPermissions bool) ([]*wmdef.Process, map[int32]*procutil.Stats) {
-	wlmProcs := createTestWLMProcesses(createTime)
-	statsByPid := createTestWLMProcessStats(wlmProcs, elevatedPermissions)
-	return wlmProcs, statsByPid
-}
-
-func createTestWLMProcesses(createTime time.Time) []*wmdef.Process {
-	nowSeconds := createTime.Unix()
-	proc1 := wlmProcessWithCreateTime(1, "git clone google.com", nowSeconds)
-	proc2 := wlmProcessWithCreateTime(2, "mine-bitcoins -all -x", nowSeconds+1)
-	proc3 := wlmProcessWithCreateTime(3, "foo --version", nowSeconds+2)
-	proc4 := wlmProcessWithCreateTime(4, "foo -bar -bim", nowSeconds+3)
-	proc5 := wlmProcessWithCreateTime(5, "datadog-agent --cfgpath datadog.conf", nowSeconds+2)
-	proc6 := wlmProcessWithCreateTime(6, "/bin/bash/usr/local/bin/cilium-agent-bpf-map-metrics.sh", nowSeconds+4)
-	return []*wmdef.Process{proc1, proc2, proc3, proc4, proc5, proc6}
-}
-
-func createTestWLMProcessStats(wlmProcs []*wmdef.Process, elevatedPermissions bool) map[int32]*procutil.Stats {
-	statsByPid := make(map[int32]*procutil.Stats, len(wlmProcs))
-	for _, wlmProc := range wlmProcs {
-		statsByPid[wlmProc.Pid] = randomProcessStats(wlmProc.CreationTime.Unix(), elevatedPermissions)
-	}
-	return statsByPid
-}
-
-func newMockClock(time time.Time) *clock.Mock {
-	c := clock.NewMock()
-	c.Set(time)
-	return c
-}
-
-func expectedRealtimeStats(wlmProcs []*wmdef.Process, statsByPid map[int32]*procutil.Stats, lastStatsByPid map[int32]*procutil.Stats, now time.Time, lastRunTime time.Time) []*model.ProcessStat {
-	expectedStats := make([]*model.ProcessStat, len(wlmProcs))
-	for i, wlmProc := range wlmProcs {
-		stats := statsByPid[wlmProc.Pid]
-		lastStats := lastStatsByPid[wlmProc.Pid]
-		expectedStat := &model.ProcessStat{
-			Pid:                    wlmProc.Pid,
-			CreateTime:             stats.CreateTime,
-			Memory:                 formatMemory(stats),
-			Cpu:                    formatCPU(stats, lastStats, procutilCPUTimeToCPUTimeStat(stats.CPUTime), procutilCPUTimeToCPUTimeStat(lastStats.CPUTime)),
-			Nice:                   stats.Nice,
-			Threads:                stats.NumThreads,
-			OpenFdCount:            stats.OpenFdCount,
-			ProcessState:           model.ProcessState(model.ProcessState_value[stats.Status]),
-			IoStat:                 formatIO(stats, lastStats.IOStat, now, lastRunTime),
-			VoluntaryCtxSwitches:   uint64(stats.CtxSwitches.Voluntary),
-			InvoluntaryCtxSwitches: uint64(stats.CtxSwitches.Involuntary),
-			//ContainerId:            wlmProc.ContainerID, // TODO: current not tested, will be done when container data is added
-		}
-		expectedStats[i] = expectedStat
-	}
-	return expectedStats
-}
-
-func TestProcessCheckRunWLM3Times(t *testing.T) {
-	for _, tc := range []struct {
-		name                string
-		elevatedPermissions bool
-		realtimeCollection  bool
-	}{
-		{
-			name:                "realtime DISABLED, elevated permissions DISABLED",
-			elevatedPermissions: false,
-			realtimeCollection:  false,
-		},
-		{
-			name:                "realtime DISABLED, elevated permissions ENABLED",
-			elevatedPermissions: true,
-			realtimeCollection:  false,
-		},
-		{
-			name:                "realtime ENABLED, elevated permissions DISABLED",
-			elevatedPermissions: false,
-			realtimeCollection:  true,
-		},
-		{
-			name:                "realtime ENABLED, elevated permissions ENABLED",
-			elevatedPermissions: true,
-			realtimeCollection:  true,
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-
-			processCheck, mockProbe, mockWLM := processCheckWithMockProbeWLM(t, tc.elevatedPermissions)
-			now := time.Now()
-			constantClock1 := clock.NewMock()
-			processCheck.timer = constantClock1
-
-			// CREATE TEST DATA
-			wlmProcs, statsByPid1 := createTestWLMProcessData(now, tc.elevatedPermissions)
-			proc1, proc2, proc3, proc4, proc5, proc6 := wlmProcs[0], wlmProcs[1], wlmProcs[2], wlmProcs[3], wlmProcs[4], wlmProcs[5]
-
-			// MOCK WLM
-			mockWLM.On("ListProcesses").Return(wlmProcs)
-
-			// MOCK PROBE FOR STATS
-			pids := make([]int32, len(wlmProcs))
-			for i, wlmProc := range wlmProcs {
-				pids[i] = wlmProc.Pid
-			}
-			mockProbe.On("StatsForPIDs", pids, mock.Anything).Return(statsByPid1, nil).Once()
-
-			// TEST FUNCTION
-			actual, err := processCheck.runWLM(0, tc.realtimeCollection)
-
-			// The first run returns nothing because processes must be observed on two consecutive runs
-			expected1 := CombinedRunResult{}
-			require.NoError(t, err)
-			assert.Equal(t, expected1, actual)
-
-			// SECOND RUN
-			constantClock2 := newMockClock(now.Add(10 * time.Second))
-			processCheck.timer = constantClock2
-			statsByPid2 := createTestWLMProcessStats(wlmProcs, tc.elevatedPermissions)
-			mockProbe.On("StatsForPIDs", pids, mock.Anything).Return(statsByPid2, nil).Once()
-
-			expected2 := []model.MessageBody{
-				&model.CollectorProc{
-					Processes: []*model.Process{modelProcFromWLMProc(proc1, statsByPid2[proc1.Pid], statsByPid1[proc1.Pid], []string{"process_context:git"}, constantClock2.Now(), constantClock1.Now())},
-					GroupSize: int32(len(wlmProcs)),
-					Info:      processCheck.hostInfo.SystemInfo,
-					Hints:     &model.CollectorProc_HintMask{HintMask: 0b1},
-				},
-				&model.CollectorProc{
-					Processes: []*model.Process{modelProcFromWLMProc(proc2, statsByPid2[proc2.Pid], statsByPid1[proc2.Pid], []string{"process_context:mine-bitcoins"}, constantClock2.Now(), constantClock1.Now())},
-					GroupSize: int32(len(wlmProcs)),
-					Info:      processCheck.hostInfo.SystemInfo,
-					Hints:     &model.CollectorProc_HintMask{HintMask: 0b1},
-				},
-				&model.CollectorProc{
-					Processes: []*model.Process{modelProcFromWLMProc(proc3, statsByPid2[proc3.Pid], statsByPid1[proc3.Pid], []string{"process_context:foo"}, constantClock2.Now(), constantClock1.Now())},
-					GroupSize: int32(len(wlmProcs)),
-					Info:      processCheck.hostInfo.SystemInfo,
-					Hints:     &model.CollectorProc_HintMask{HintMask: 0b1},
-				},
-				&model.CollectorProc{
-					Processes: []*model.Process{modelProcFromWLMProc(proc4, statsByPid2[proc4.Pid], statsByPid1[proc4.Pid], []string{"process_context:foo"}, constantClock2.Now(), constantClock1.Now())},
-					GroupSize: int32(len(wlmProcs)),
-					Info:      processCheck.hostInfo.SystemInfo,
-					Hints:     &model.CollectorProc_HintMask{HintMask: 0b1},
-				},
-				&model.CollectorProc{
-					Processes: []*model.Process{modelProcFromWLMProc(proc5, statsByPid2[proc5.Pid], statsByPid1[proc5.Pid], []string{"process_context:datadog-agent"}, constantClock2.Now(), constantClock1.Now())},
-					GroupSize: int32(len(wlmProcs)),
-					Info:      processCheck.hostInfo.SystemInfo,
-					Hints:     &model.CollectorProc_HintMask{HintMask: 0b1},
-				},
-				&model.CollectorProc{
-					Processes: []*model.Process{modelProcFromWLMProc(proc6, statsByPid2[proc6.Pid], statsByPid1[proc6.Pid], []string{"process_context:cilium-agent-bpf-map-metrics"}, constantClock2.Now(), constantClock1.Now())},
-					GroupSize: int32(len(wlmProcs)),
-					Info:      processCheck.hostInfo.SystemInfo,
-					Hints:     &model.CollectorProc_HintMask{HintMask: 0b1},
-				},
-			}
-			actual, err = processCheck.runWLM(0, tc.realtimeCollection)
-			require.NoError(t, err)
-			assert.ElementsMatch(t, expected2, actual.Payloads())
-			// realtime collection
-			if tc.realtimeCollection {
-				expectedStats := expectedRealtimeStats(wlmProcs, statsByPid2, statsByPid1, constantClock2.Now(), constantClock1.Now())
-				require.Len(t, actual.RealtimePayloads(), 1)
-				rt := actual.RealtimePayloads()[0].(*model.CollectorRealTime)
-				assert.ElementsMatch(t, expectedStats, rt.Stats)
-				assert.Equal(t, int32(1), rt.GroupSize)
-				assert.Equal(t, int32(len(processCheck.hostInfo.SystemInfo.Cpus)), rt.NumCpus)
-			} else {
-				assert.Nil(t, actual.RealtimePayloads())
-			}
-
-			// THIRD RUN
-			constantClock3 := newMockClock(now.Add(20 * time.Second))
-			processCheck.timer = constantClock3
-			statsByPid3 := createTestWLMProcessStats(wlmProcs, tc.elevatedPermissions)
-			mockProbe.On("StatsForPIDs", pids, mock.Anything).Return(statsByPid3, nil).Once()
-
-			expected3 := []model.MessageBody{
-				&model.CollectorProc{
-					Processes: []*model.Process{modelProcFromWLMProc(proc1, statsByPid3[proc1.Pid], statsByPid2[proc1.Pid], []string{"process_context:git"}, constantClock3.Now(), constantClock2.Now())},
-					GroupSize: int32(len(wlmProcs)),
-					Info:      processCheck.hostInfo.SystemInfo,
-					Hints:     &model.CollectorProc_HintMask{HintMask: 0b0},
-				},
-				&model.CollectorProc{
-					Processes: []*model.Process{modelProcFromWLMProc(proc2, statsByPid3[proc2.Pid], statsByPid2[proc2.Pid], []string{"process_context:mine-bitcoins"}, constantClock3.Now(), constantClock2.Now())},
-					GroupSize: int32(len(wlmProcs)),
-					Info:      processCheck.hostInfo.SystemInfo,
-					Hints:     &model.CollectorProc_HintMask{HintMask: 0b0},
-				},
-				&model.CollectorProc{
-					Processes: []*model.Process{modelProcFromWLMProc(proc3, statsByPid3[proc3.Pid], statsByPid2[proc3.Pid], []string{"process_context:foo"}, constantClock3.Now(), constantClock2.Now())},
-					GroupSize: int32(len(wlmProcs)),
-					Info:      processCheck.hostInfo.SystemInfo,
-					Hints:     &model.CollectorProc_HintMask{HintMask: 0b0},
-				},
-				&model.CollectorProc{
-					Processes: []*model.Process{modelProcFromWLMProc(proc4, statsByPid3[proc4.Pid], statsByPid2[proc4.Pid], []string{"process_context:foo"}, constantClock3.Now(), constantClock2.Now())},
-					GroupSize: int32(len(wlmProcs)),
-					Info:      processCheck.hostInfo.SystemInfo,
-					Hints:     &model.CollectorProc_HintMask{HintMask: 0b0},
-				},
-				&model.CollectorProc{
-					Processes: []*model.Process{modelProcFromWLMProc(proc5, statsByPid3[proc5.Pid], statsByPid2[proc5.Pid], []string{"process_context:datadog-agent"}, constantClock3.Now(), constantClock2.Now())},
-					GroupSize: int32(len(wlmProcs)),
-					Info:      processCheck.hostInfo.SystemInfo,
-					Hints:     &model.CollectorProc_HintMask{HintMask: 0b0},
-				},
-				&model.CollectorProc{
-					Processes: []*model.Process{modelProcFromWLMProc(proc6, statsByPid3[proc6.Pid], statsByPid2[proc6.Pid], []string{"process_context:cilium-agent-bpf-map-metrics"}, constantClock3.Now(), constantClock2.Now())},
-					GroupSize: int32(len(wlmProcs)),
-					Info:      processCheck.hostInfo.SystemInfo,
-					Hints:     &model.CollectorProc_HintMask{HintMask: 0b0},
-				},
-			}
-			actual, err = processCheck.runWLM(0, tc.realtimeCollection)
-			require.NoError(t, err)
-			assert.ElementsMatch(t, expected3, actual.Payloads())
-			if tc.realtimeCollection {
-				expectedStats := expectedRealtimeStats(wlmProcs, statsByPid3, statsByPid2, constantClock3.Now(), constantClock2.Now())
-				require.Len(t, actual.RealtimePayloads(), 1)
-				rt := actual.RealtimePayloads()[0].(*model.CollectorRealTime)
-				assert.ElementsMatch(t, expectedStats, rt.Stats)
-				assert.Equal(t, int32(1), rt.GroupSize)
-				assert.Equal(t, int32(len(processCheck.hostInfo.SystemInfo.Cpus)), rt.NumCpus)
-			} else {
-				assert.Nil(t, actual.RealtimePayloads())
-			}
-		})
-	}
-}
-
-func TestProcessCheckChunkingWLM(t *testing.T) {
-	for _, tc := range []struct {
-		name                  string
-		noChunking            bool
-		expectedPayloadLength int
-		elevatedPermissions   bool
-	}{
-		{
-			name:                  "Chunking ENABLED, elevated permissions DISABLED",
-			noChunking:            false,
-			expectedPayloadLength: 6,
-			elevatedPermissions:   false,
-		},
-		{
-			name:                  "Chunking DISABLED, elevated permissions DISABLED",
-			noChunking:            true,
-			expectedPayloadLength: 1,
-			elevatedPermissions:   false,
-		},
-		{
-			name:                  "Chunking ENABLED, elevated permissions ENABLED",
-			noChunking:            false,
-			expectedPayloadLength: 6,
-			elevatedPermissions:   true,
-		},
-		{
-			name:                  "Chunking DISABLED, elevated permissions ENABLED",
-			noChunking:            true,
-			expectedPayloadLength: 1,
-			elevatedPermissions:   true,
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			// mock processes
-			now := time.Now()
-			processCheck, mockProbe, mockWLM := processCheckWithMockProbeWLM(&testing.T{}, tc.elevatedPermissions)
-			wlmProcs, statsByPid := createTestWLMProcessData(now, tc.elevatedPermissions)
-
-			// MOCK WLM
-			mockWLM.On("ListProcesses").Return(wlmProcs)
-
-			// MOCK PROBE FOR STATS
-			pids := make([]int32, len(wlmProcs))
-			for i, wlmProc := range wlmProcs {
-				pids[i] = wlmProc.Pid
-			}
-			mockProbe.On("StatsForPIDs", pids, mock.Anything).Return(statsByPid, nil)
-
-			// Set small chunk size to force chunking behavior
-			processCheck.maxBatchBytes = 0
-			processCheck.maxBatchSize = 0
-
-			// Test second check runs without error and has correct number of chunks
-			processCheck.Run(testGroupID(0), getChunkingOption(tc.noChunking))
-			actual, err := processCheck.Run(testGroupID(0), getChunkingOption(tc.noChunking))
-			require.NoError(t, err)
-			assert.Len(t, actual.Payloads(), tc.expectedPayloadLength)
-		})
-	}
-}
-
-func TestProcessCheckZombieConfig(t *testing.T) {
-	for _, tc := range []struct {
-		name                  string
-		ignoreZombieProcesses bool
-		elevatedPermissions   bool
-		realtimeCollection    bool
-	}{
-		{
-			name:                  "ignore zombie processes ENABLED, realtime ENABLED, elevated permissions ENABLED",
-			ignoreZombieProcesses: true,
-			realtimeCollection:    true,
-			elevatedPermissions:   true,
-		},
-		{
-			name:                  "Ignore zombie processes ENABLED, realtime ENABLED, elevated permissions DISABLED",
-			ignoreZombieProcesses: true,
-			realtimeCollection:    true,
-			elevatedPermissions:   false,
-		},
-		{
-			name:                  "ignore zombie processes ENABLED, realtime DISABLED, elevated permissions ENABLED",
-			ignoreZombieProcesses: true,
-			realtimeCollection:    false,
-			elevatedPermissions:   true,
-		},
-		{
-			name:                  "Ignore zombie processes ENABLED, realtime DISABLED, elevated permissions DISABLED",
-			ignoreZombieProcesses: true,
-			realtimeCollection:    false,
-			elevatedPermissions:   false,
-		},
-		{
-			name:                  "ignore zombie processes DISABLED, realtime ENABLED, elevated permissions ENABLED",
-			ignoreZombieProcesses: false,
-			realtimeCollection:    true,
-			elevatedPermissions:   true,
-		},
-		{
-			name:                  "Ignore zombie processes DISABLED, realtime ENABLED, elevated permissions DISABLED",
-			ignoreZombieProcesses: false,
-			realtimeCollection:    true,
-			elevatedPermissions:   false,
-		},
-		{
-			name:                  "ignore zombie processes DISABLED, realtime DISABLED, elevated permissions ENABLED",
-			ignoreZombieProcesses: false,
-			realtimeCollection:    false,
-			elevatedPermissions:   true,
-		},
-		{
-			name:                  "Ignore zombie processes DISABLED, realtime DISABLED, elevated permissions DISABLED",
-			ignoreZombieProcesses: false,
-			realtimeCollection:    false,
-			elevatedPermissions:   false,
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			// mock processes
-			processCheck, mockProbe, mockWLM := processCheckWithMockProbeWLM(&testing.T{}, tc.elevatedPermissions)
-			now := time.Now()
-			constantClock1 := newMockClock(now)
-			processCheck.timer = constantClock1
-			cfg := configmock.New(t)
-			cfg.SetWithoutSource("process_config.ignore_zombie_processes", tc.ignoreZombieProcesses)
-			processCheck.config = cfg
-			processCheck.ignoreZombieProcesses = processCheck.config.GetBool(configIgnoreZombies)
-
-			// TEST DATA
-			wlmProcs, statsByPid1 := createTestWLMProcessData(now, tc.elevatedPermissions)
-			proc1, proc2, proc3, proc4, proc5, proc6 := wlmProcs[0], wlmProcs[1], wlmProcs[2], wlmProcs[3], wlmProcs[4], wlmProcs[5]
-			// ENSURE 2 PROCESSES ARE NOT ZOMBIES
-			statsByPid1[proc1.Pid].Status = "R"
-			statsByPid1[proc2.Pid].Status = "R"
-
-			// SET 4 PROCESSES TO ZOMBIE STATUS
-			statsByPid1[proc3.Pid].Status = "Z"
-			statsByPid1[proc4.Pid].Status = "Z"
-			statsByPid1[proc5.Pid].Status = "Z"
-			statsByPid1[proc6.Pid].Status = "Z"
-
-			// MOCK WLM
-			mockWLM.On("ListProcesses").Return(wlmProcs)
-
-			// MOCK PROBE FOR STATS
-			pids := make([]int32, len(wlmProcs))
-			for i, wlmProc := range wlmProcs {
-				pids[i] = wlmProc.Pid
-			}
-			mockProbe.On("StatsForPIDs", pids, mock.Anything).Return(statsByPid1, nil).Once()
-
-			// TEST FUNCTION
-			actual, err := processCheck.runWLM(0, tc.realtimeCollection)
-
-			// The first run returns nothing because processes must be observed on two consecutive runs
-			expected1 := CombinedRunResult{}
-			require.NoError(t, err)
-			assert.Equal(t, expected1, actual)
-			// creation times should be the same
-			for _, wlmProc := range wlmProcs {
-				assert.Equal(t, wlmProc.CreationTime.Unix(), statsByPid1[wlmProc.Pid].CreateTime)
-			}
-
-			// SECOND RUN
-			constantClock2 := newMockClock(now.Add(10 * time.Second))
-			processCheck.timer = constantClock2
-			statsByPid2 := createTestWLMProcessStats(wlmProcs, tc.elevatedPermissions)
-			// ENSURE 2 PROCESSES ARE NOT ZOMBIES
-			statsByPid2[proc1.Pid].Status = "R"
-			statsByPid2[proc2.Pid].Status = "R"
-
-			// SET 4 PROCESSES TO ZOMBIE STATUS
-			statsByPid2[proc3.Pid].Status = "Z"
-			statsByPid2[proc4.Pid].Status = "Z"
-			statsByPid2[proc5.Pid].Status = "Z"
-			statsByPid2[proc6.Pid].Status = "Z"
-			// creation times should be the same
-			for _, wlmProc := range wlmProcs {
-				assert.Equal(t, wlmProc.CreationTime.Unix(), statsByPid2[wlmProc.Pid].CreateTime)
-			}
-			mockProbe.On("StatsForPIDs", pids, mock.Anything).Return(statsByPid2, nil).Once()
-
-			var expected2 []model.MessageBody
-			if tc.ignoreZombieProcesses {
-				expected2 = []model.MessageBody{
-					&model.CollectorProc{
-						Processes: []*model.Process{modelProcFromWLMProc(proc1, statsByPid2[proc1.Pid], statsByPid1[proc1.Pid], []string{"process_context:git"}, constantClock2.Now(), constantClock1.Now())},
-						GroupSize: int32(2),
-						Info:      processCheck.hostInfo.SystemInfo,
-						Hints:     &model.CollectorProc_HintMask{HintMask: 0b1},
-					},
-					&model.CollectorProc{
-						Processes: []*model.Process{modelProcFromWLMProc(proc2, statsByPid2[proc2.Pid], statsByPid1[proc2.Pid], []string{"process_context:mine-bitcoins"}, constantClock2.Now(), constantClock1.Now())},
-						GroupSize: int32(2),
-						Info:      processCheck.hostInfo.SystemInfo,
-						Hints:     &model.CollectorProc_HintMask{HintMask: 0b1},
-					},
-				}
-			} else {
-				expected2 = []model.MessageBody{
-					&model.CollectorProc{
-						Processes: []*model.Process{modelProcFromWLMProc(proc1, statsByPid2[proc1.Pid], statsByPid1[proc1.Pid], []string{"process_context:git"}, constantClock2.Now(), constantClock1.Now())},
-						GroupSize: int32(len(wlmProcs)),
-						Info:      processCheck.hostInfo.SystemInfo,
-						Hints:     &model.CollectorProc_HintMask{HintMask: 0b1},
-					},
-					&model.CollectorProc{
-						Processes: []*model.Process{modelProcFromWLMProc(proc2, statsByPid2[proc2.Pid], statsByPid1[proc2.Pid], []string{"process_context:mine-bitcoins"}, constantClock2.Now(), constantClock1.Now())},
-						GroupSize: int32(len(wlmProcs)),
-						Info:      processCheck.hostInfo.SystemInfo,
-						Hints:     &model.CollectorProc_HintMask{HintMask: 0b1},
-					},
-					&model.CollectorProc{
-						Processes: []*model.Process{modelProcFromWLMProc(proc3, statsByPid2[proc3.Pid], statsByPid1[proc3.Pid], []string{"process_context:foo"}, constantClock2.Now(), constantClock1.Now())},
-						GroupSize: int32(len(wlmProcs)),
-						Info:      processCheck.hostInfo.SystemInfo,
-						Hints:     &model.CollectorProc_HintMask{HintMask: 0b1},
-					},
-					&model.CollectorProc{
-						Processes: []*model.Process{modelProcFromWLMProc(proc4, statsByPid2[proc4.Pid], statsByPid1[proc4.Pid], []string{"process_context:foo"}, constantClock2.Now(), constantClock1.Now())},
-						GroupSize: int32(len(wlmProcs)),
-						Info:      processCheck.hostInfo.SystemInfo,
-						Hints:     &model.CollectorProc_HintMask{HintMask: 0b1},
-					},
-					&model.CollectorProc{
-						Processes: []*model.Process{modelProcFromWLMProc(proc5, statsByPid2[proc5.Pid], statsByPid1[proc5.Pid], []string{"process_context:datadog-agent"}, constantClock2.Now(), constantClock1.Now())},
-						GroupSize: int32(len(wlmProcs)),
-						Info:      processCheck.hostInfo.SystemInfo,
-						Hints:     &model.CollectorProc_HintMask{HintMask: 0b1},
-					},
-					&model.CollectorProc{
-						Processes: []*model.Process{modelProcFromWLMProc(proc6, statsByPid2[proc6.Pid], statsByPid1[proc6.Pid], []string{"process_context:cilium-agent-bpf-map-metrics"}, constantClock2.Now(), constantClock1.Now())},
-						GroupSize: int32(len(wlmProcs)),
-						Info:      processCheck.hostInfo.SystemInfo,
-						Hints:     &model.CollectorProc_HintMask{HintMask: 0b1},
-					},
-				}
-			}
-			actual, err = processCheck.runWLM(0, tc.realtimeCollection)
-			require.NoError(t, err)
-			assert.ElementsMatch(t, expected2, actual.Payloads())
-			// realtime collection
-			if tc.realtimeCollection {
-				expectedStats := expectedRealtimeStats(wlmProcs, statsByPid2, statsByPid1, constantClock2.Now(), constantClock1.Now())
-				require.Len(t, actual.RealtimePayloads(), 1)
-				rt := actual.RealtimePayloads()[0].(*model.CollectorRealTime)
-				assert.ElementsMatch(t, expectedStats, rt.Stats)
-				assert.Equal(t, int32(1), rt.GroupSize)
-				assert.Equal(t, int32(len(processCheck.hostInfo.SystemInfo.Cpus)), rt.NumCpus)
-			} else {
-				assert.Nil(t, actual.RealtimePayloads())
-			}
-		})
-	}
-}
-
-// BenchmarkProcessCheckRUNWLM does not run with the typical dda inv test, you will have to edit the invoke task
-// to include the bench flag in tasks/gotest.py, here are the previous results on a m3 max
-// BenchmarkProcessCheckRunWLM-16             30969             37086 ns/op           27917 B/op        301 allocs/op
-// BenchmarkProcessCheck-16                   41541             28592 ns/op           22013 B/op        243 allocs/op
-func BenchmarkProcessCheckRunWLM(b *testing.B) {
-	elevatedPermissions := true
-	processCheck, mockProbe, mockWLM := processCheckWithMockProbeWLM(&testing.T{}, elevatedPermissions)
-	wlmProcs, statsByPid := createTestWLMProcessData(time.Now(), elevatedPermissions)
-
-	// MOCK WLM
-	mockWLM.On("ListProcesses").Return(wlmProcs)
-
-	// MOCK PROBE FOR STATS
-	pids := make([]int32, len(wlmProcs))
-	for i, wlmProc := range wlmProcs {
-		pids[i] = wlmProc.Pid
-	}
-	mockProbe.On("StatsForPIDs", pids, mock.Anything).Return(statsByPid, nil)
-
-	for n := 0; n < b.N; n++ {
-		_, err := processCheck.runWLM(0, false)
-		require.NoError(b, err)
 	}
 }
