+++ pkg/process/checks/process_test.go
@@ -14,6 +14,7 @@ import (
 
 	model "github.com/DataDog/agent-payload/v5/process"
 	"github.com/DataDog/datadog-go/v5/statsd"
+	"github.com/benbjohnson/clock"
 	"github.com/shirou/gopsutil/v4/cpu"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -71,6 +72,7 @@ func processCheckWithMockProbe(t *testing.T) (*ProcessCheck, *mocks.Probe) {
 		extractors:        []metadata.Extractor{serviceExtractor},
 		gpuSubscriber:     mockGpuSubscriber,
 		statsd:            &statsd.NoOpClient{},
+		clock:             clock.NewMock(),
 	}, probe
 }
 
@@ -470,25 +472,6 @@ func TestProcessWithNoCommandline(t *testing.T) {
 	proc := procs[""][0]
 	assert.Equal(t, procMap[1].Exe, proc.Command.Exe)
 	assert.Empty(t, proc.Command.Args)
-}
-
-func BenchmarkProcessCheck(b *testing.B) {
-	processCheck, probe := processCheckWithMockProbe(&testing.T{})
-
-	now := time.Now().Unix()
-	proc1 := makeProcessWithCreateTime(1, "git clone google.com", now)
-	proc2 := makeProcessWithCreateTime(2, "mine-bitcoins -all -x", now+1)
-	proc3 := makeProcessWithCreateTime(3, "foo --version", now+2)
-	proc4 := makeProcessWithCreateTime(4, "foo -bar -bim", now+3)
-	proc5 := makeProcessWithCreateTime(5, "datadog-process-agent --cfgpath datadog.conf", now+2)
-	processesByPid := map[int32]*procutil.Process{1: proc1, 2: proc2, 3: proc3, 4: proc4, 5: proc5}
-
-	probe.On("ProcessesByPID", mock.Anything, mock.Anything).Return(processesByPid, nil)
-
-	for n := 0; n < b.N; n++ {
-		_, err := processCheck.run(0, false)
-		require.NoError(b, err)
-	}
 }
 
 func TestProcessCheckZombieToggleFalse(t *testing.T) {
